rewrite of esp_main.ino
Ready to connect to the atm_main.ino. Communications passed to the esp by the atmega are used to communicate a parameter datum, send a post consisting of a collection of parameter data, or query the timestring from AWS.

Acceptable communications to the ESP (see lines 113 - 139) are of the format [t 23.45x] where 't' is the command character (in this case, set the temperature value), followed by a space, then up to 22 characters of content, and terminated with an x. Other command strings in lines 113-139 are for setting the humidity, battery voltage, ozone, nitrogen dioxide, sulfur dioxide, and hydrogen sulfide values for a post to AWS. All of these parameters must be given a value before a post can be made successfully. The command 'px' is used to trigger a post action; the command 'dx' can also be used to query AWS for the timestamp (only after the wifi has successfully connected, of course).

As currently coded the esp will only return a value to the atmega after a post attempt or a timestamp query. The code "#S" is returned for a successful post, and the codes "#E1" through "#E5" are used to indicate various errors. 

Feel free to edit or add comments as needed; I'll try to get some better commenting up this week.